requirements.txt
fastapi
uvicorn
motor
pydantic[email]
passlib[bcrypt]
python-jose
python-dotenv
httpx
pytest
